gralloc_headers: Add legacy gralloc headers and selector
Add legacy gralloc headers that are necessary for the correct
operation of legacy gralloc and other system components that
depend on these headers.

Also based on the fact that we cannot use conditions in blueprints,
add a gralloc selector handler which was written on Go and was
added to this repo to correctly select the headers we need.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>